Functional test for creating and revoking exchanges
* Moving E501 ignoring flag in the config file
* Excluding conftest.py files from flake8 analysis, subject to F401

Setup functional tests with behave

cognito api integration and refacto

switches from email to sub for linking user identity to exchanges

refacto

Fix flake8 style issues

+++ app/hallebarde/revoke_exchange.py
@@ -1,22 +1,34 @@
 import logging
+from http import HTTPStatus
 
 from hallebarde.infrastructure import exchange_repository
 from hallebarde.infrastructure import file_repository
-from hallebarde.infrastructure.event_parser import extract_from_headers
+from hallebarde.infrastructure import event_parser
+from hallebarde import config
 
 logger = logging.getLogger()
 logger.setLevel(logging.INFO)
 
 
 def handle(event: dict, context: dict) -> dict:
-    identifier = extract_from_headers('exchange_identifier', event)
+    jwt_token = event_parser.extract_from_headers(config.AUTHORIZATION_HEADER, event)
+    event_sub = event_parser.extract_sub_from_jwt(jwt_token)
+    identifier = event_parser.extract_from_headers('exchange_identifier', event)
 
-    file_repository.delete_files(identifier)
-    exchange_repository.delete(identifier)
+    exchange = exchange_repository.get(identifier)
 
+    if exchange.sub != event_sub:
+        return _generate_response("Forbidden", HTTPStatus.FORBIDDEN)
+    else:
+        file_repository.delete_files(identifier)
+        exchange_repository.delete(identifier)
+        return _generate_response("Exchange revoked", HTTPStatus.OK)
+
+
+def _generate_response(body: str, status_code: int):
     return {
         "isBase64Encoded": False,
-        "body": None,
+        "body": body,
         "headers": None,
-        "statusCode": 200
+        "statusCode": status_code
     }
